+++ apps/vscode/src/test/examples/generated_snapshots/vdoc/magics-python.py
@@ -1,0 +1,25 @@
+# type: ignore
+# flake8: noqa
+#
+#
+#
+#
+#
+#
+#
+#
+import pandas as pd
+
+#
+#
+#
+#
+x = 1
+#
+#
+#
+#
+#
+#
+#
+#